Block access to auth page and display page empty page when recipes list is empty

+++ auth/recipes.php
@@ -1,22 +1,11 @@
-<?php 
-/**
- * Connexion à la base de données
- */
+<?php
+session_start();
+require_once("../inc/functions.php");
 require_once("../conn_db/conn.php");
 
-/***
- * Titre de la page
- */
-$titlePage="Recettes";
+block_access();
 
-/**
- * Insertion du header
- */
-require_once('../inc/header.php');
-
-/***
- * Recuoperations des données
- */
+$titlePage = "Recettes";
 $userId = (int) getLoggedUserInfo('userId');
 
 // Requette SQL de recuperation de la liste des requettes
@@ -29,23 +18,36 @@ $recipesStatement->execute([
 ]);
 // recuperation des elements de la requettes
 $recipes = $recipesStatement->fetchAll(PDO::FETCH_ASSOC);
+require_once('../inc/header.php');
 ?>
-<h3 class="mt-3 fw-bolder"> <?= count($recipes) ?> Recettes</h3>
+<h3 class="mt-3 fw-bolder"> <?= count($recipes) ?> Recette<?= count($recipes) > 1 ? "s" : "" ?></h3>
 <hr>
 <!-- Parcours et affichage des recettes -->
-<?php foreach ($recipes as $recipe) : extract($recipe) ?>
-    <article class="mb-3 bg-light rounded p-3 border border-dark">
-    <h3 class="fw-bolder"><a target="_blank" class="text-dark text-decoration-none" href="recipe.php?id=<?= $recipe['recipe_id'] ?>"><?php echo $recipe['title']; ?></a></h3>
-        <div><?php echo $recipe['description']; ?></div>
-        <div class="d-flex mt-3">
-            <a href="recipe_edit.php?id=<?= $recipe_id ?>" class="btn btn-success me-2">Modifier</a>
-            <a 
-            onclick="return confirm('Voulez vous vraiment supprimer cette recette ?')"
-            
-            href="recipe_delete.php?id=<?= $recipe_id ?>"  class="btn btn-danger">Supprimer</a>
+<?php if (count($recipes) == 0) : ?>
+    <div class="card-body p-0 ">
+        <div class="d-flex flex-column p-5 w-50 text-center  mx-auto">
+            <i class="fa fa-user fs-1 text-app" aria-hidden="true"></i>
+            <h3 class="text-app-secondary fw-bold">Vous n'avez creer aucune recette</h3>
+            <p class="text-muted">Tous vos recettes seront sauvegardées ici pour que vous puissez les consultertout moment.
+            </p>
+            <a href="recipe_form.php" class="w-50 mx-auto btn-lg btn btn-dark">
+                <i class="fa fa-plus"></i> Creer une recette</a>
         </div>
-    </article>
-<?php endforeach ?>
-</div>
-<?php
-require_once("../inc/footer.php");
+    </div>
+        <?php else :
+        foreach ($recipes as $recipe) : extract($recipe) ?>
+            <article class="mb-3 bg-light rounded p-3 border border-dark">
+                <h3 class="fw-bolder"><a target="_blank" class="text-dark text-decoration-none" href="recipe.php?id=<?= $recipe['recipe_id'] ?>"><?php echo $recipe['title']; ?></a></h3>
+                <div><?php echo $recipe['description']; ?></div>
+                <div class="d-flex mt-3">
+                    <a href="recipe_edit.php?id=<?= $recipe_id ?>" class="btn btn-success me-2">Modifier</a>
+                    <a onclick="return confirm('Voulez vous vraiment supprimer cette recette ?')" href="recipe_delete.php?id=<?= $recipe_id ?>" class="btn btn-danger">Supprimer</a>
+                </div>
+            </article>
+        <?php 
+    endforeach;
+    endif;
+    ?>
+    </div>
+    <?php
+    require_once("../inc/footer.php");
